feat: add JavaScript API
