+++ packages/eslint-rule-explorer/bin/eslint-rule-explorer.mjs
@@ -1,5 +1,5 @@
 #!/usr/bin/env node
 // @ts-check
-import { run } from "../dist/index.js";
+import { run } from "../dist/cli.js";
 
 await run(process.argv.slice(2));
