+++ packages/eslint-rule-explorer/src/index.ts
@@ -1,80 +1,57 @@
-import type { Args } from "gunshi";
-
 import { resolveFlatConfig } from "@sushichan044/eslint-config-array-resolver";
-import { cli, define } from "gunshi";
 import { cwd } from "node:process";
 
-import {
-  description as pkgDescription,
-  name as pkgName,
-  version as pkgVersion,
-} from "../package.json";
-import { formatRuleInfo } from "./format";
-import { searchRule } from "./search";
+import type { RuleFindOptions } from "./find";
+import type { RuleMetaData } from "./types";
 
-const mainCmd = define({
-  args: {
-    exact: {
-      default: false,
-      description: "Search for the exact rule name",
-      type: "boolean",
-    },
-    json: {
-      default: false,
-      description: "Output as JSON",
-      type: "boolean",
-    },
-    root: {
-      description:
-        "Root directory of the project (default: current working directory)",
-      type: "string",
-    },
-    rule: {
-      description:
-        "Rule name to search for (e.g., 'no-unused-vars', '@typescript-eslint/no-explicit-any')",
-      type: "positional",
-    },
-  } as const satisfies Args,
-  name: "root",
-  run: async (c) => {
-    const { exact: searchExact, json: printAsJSON, rule: ruleName } = c.values;
+import { findRule } from "./find";
 
-    const rootDir = c.values.root ?? cwd();
-    const eslintConfig = await resolveFlatConfig(rootDir, {
-      // unnecessary stdout will break JSON pipeline like jq
-      suppressOutput: printAsJSON,
-    });
+interface RuleSearchOptions extends RuleFindOptions {
+  /**
+   * Root directory of the project
+   *
+   * @default `process.cwd()`
+   */
+  rootDir?: string;
 
-    const searchInput = {
+  /**
+   * Suppress any console output during ESLint configuration resolution.
+   *
+   * Useful when using the output in JSON pipelines where extra logs would break parsing.
+   */
+  suppressOutput?: boolean;
+}
+
+interface RuleSearchResult {
+  info: RuleMetaData;
+  name: string;
+}
+
+/**
+ * Search for ESLint rules in the project configuration.
+ *
+ * @param ruleName - Rule name to search for (e.g., `no-unused-vars`, `@typescript-eslint/no-explicit-any`)
+ * @param options - Search options
+ * @returns Promise resolving to an array of matching rules
+ * @throws Error if no rules are found or configuration is invalid
+ */
+export async function searchESLintRule(
+  ruleName: string,
+  options: RuleSearchOptions = {},
+): Promise<RuleSearchResult[]> {
+  const { rootDir = cwd(), suppressOutput, ...searchOptions } = options;
+
+  const eslintConfig = await resolveFlatConfig(rootDir, {
+    ...(suppressOutput != null && { suppressOutput }),
+  });
+
+  const searchResult = findRule(
+    {
       config: eslintConfig,
       ruleName,
-    };
+    },
+    searchOptions,
+  );
 
-    const searchResult = searchRule(searchInput, {
-      strategy: searchExact ? "exact" : "fuzzy",
-    });
-
-    if (printAsJSON) {
-      console.log(JSON.stringify(searchResult.rules, null, 2));
-      return;
-    }
-
-    if (!searchResult.found) {
-      console.log(`No rules found for input "${ruleName}"`);
-      return;
-    }
-
-    for (const { info, name } of searchResult.rules) {
-      const formattedInfo = formatRuleInfo(info);
-      console.log(`### ${name}\n\n${formattedInfo}\n`);
-    }
-  },
-});
-
-export const run = async (argv: string[]): Promise<string | undefined> =>
-  cli(argv, mainCmd, {
-    description: pkgDescription,
-    name: pkgName,
-    renderHeader: null, // disable header rendering
-    version: pkgVersion,
-  });
+  return searchResult.rules;
+}
